+++ hairsalon_app/report_view/routes.py
@@ -1,7 +1,6 @@
 from flask import Blueprint, flash, redirect, render_template, url_for
 from flask_login import current_user, login_required
 from hairsalon_app.report_view.forms import ReportForm, ReportEdit
-from hairsalon_app.appointment_view.appointment import Appointment
 from hairsalon_app.qdb.database import db
 
 
@@ -11,6 +10,7 @@ report_bp = Blueprint('report_bp', __name__, template_folder='templates', static
 
 #route to create report
 @report_bp.route('/report/<int:appointment_id>/', methods=['POST', 'GET'])
+@login_required 
 def create_report(appointment_id):
     form = ReportForm() 
     if form.validate_on_submit():
@@ -20,21 +20,14 @@ def create_report(appointment_id):
     flash('Invalid Inputs.' 'error')
     return render_template('report.html', form=form)
 
-#route for all reports
-@report_bp.route("/all_reports/", methods=['GET'])
-def all_reports(): #the id is the one for the note
-    #get reports from db
-    all_reports = db.get_all_reports()
-     
-    if (len(all_reports)!= 0):
-        return render_template("all_reports.html", context = all_reports)
-        
-    return redirect(url_for("report_bp.create_report"))
-
 #route to create report
 @report_bp.route('/edit_report/<int:report_id>', methods=['POST', 'GET'])
+@login_required 
 def edit_report(report_id):
-    report = db.get_report_by_id(report_id=report_id)
+    report = db.reports_cond(cond=f"WHERE report_id = {report_id} AND user_id = {current_user.user_id}")[0]
+    if not report:
+        flash('Report not found, or not your report.' 'info')
+        return redirect(url_for('appointment_bp.specific_appointment', appointment_id=report.appointment_id))
     form = ReportEdit() 
     if form.validate_on_submit():
         db.edit_report(report_id, form.client_report.data, form.professional_report.data)
@@ -44,7 +37,12 @@ def edit_report(report_id):
     return render_template('edit_report.html', form=form, report=report)
 
 @report_bp.route('/delete_report/<int:report_id>/<int:appointment_id>/', methods=['POST', 'GET'])
+@login_required 
 def delete_report(report_id, appointment_id):
+    report = db.reports_cond(cond=f"WHERE report_id = {report_id} AND user_id = {current_user.user_id}")[0]
+    if not report:
+        flash('Report not found, or not your report.' 'info')
+        return redirect(url_for('appointment_bp.specific_appointment', appointment_id=report.appointment_id))
     db.delete_report(report_id)
     flash('Report deleted','success')
     return redirect(url_for('appointment_bp.specific_appointment', appointment_id=appointment_id))
